requirements.txt
django==1.9.2
django-jsonfield==0.9.19
djangorestframework==3.3.1
ipython==4.0.1
pytz==2015.07
requests==2.9.1
mysqlclient==1.3.7
python_dateutil==2.4.2
gunicorn==19.4.5
django-configurations==1.0
django-crispy-forms==1.6.0
django-filter==0.12.0
django-sendfile==0.3.10
psycopg2==2.6.1
celery==3.1.20
virtualenv==14.0.6
ldap3==2.2.1

watchdog==0.8.3    # supervisor
meld3==1.0.2    # supervisor

XlsxWriter==0.8.4    # adltrack

